Исправлена валидация
